Create jenkins.sh2
Jenkins.sh DID NOT install Jenkins but rather DOCKER. Hence i created this file with the code for installing for Jenkins. I wasn't sure if it was an error so i just created a new file so as not to tamper with the code in the other file. I named this one jenkins.sh2, to avoid confusing it with the original file.
The code has been tested and it works as desired.